# Conflicts:
#	api/src/main/java/burgermap/service/MemberLookupService.java

+++ api/src/main/java/burgermap/service/ReviewService.java
@@ -1,7 +1,11 @@
 package burgermap.service;
 
 import burgermap.entity.Food;
+import burgermap.entity.Member;
 import burgermap.entity.Review;
+import burgermap.exception.food.FoodNotExistException;
+import burgermap.exception.member.MemberNotExistException;
+import burgermap.exception.review.NotReviewAuthorException;
 import burgermap.repository.ReviewRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -9,6 +13,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
 
 @Slf4j
 @Service
@@ -24,10 +30,17 @@ public class ReviewService {
     @Transactional
     public void addReview(Review review, Long memberId, Long foodId) {
         // TODO: 회원이 CUSTOMER 타입인지 확인, 아닌 경우 예외
-        Food food = foodLookupService.findByFoodId(foodId);
+        CompletableFuture<Optional<Food>> foodAsync = foodLookupService.findByFoodIdAsync(foodId);
+        CompletableFuture<Optional<Member>> memberAsync = memberLookupService.findByMemberIdAsync(memberId);
+
+        CompletableFuture.allOf(foodAsync, memberAsync).join();
+
+        Food food = foodAsync.join().orElseThrow(() -> new FoodNotExistException(foodId));
+        Member member = memberAsync.join().orElseThrow(() -> new MemberNotExistException(memberId));
+
         review.setFood(food);
         review.setStore(food.getStore());
-        review.setMember(memberLookupService.findByMemberId(memberId));
+        review.setMember(member);
         repository.save(review);
         log.debug("review added: {}", review);
     }
@@ -38,10 +51,13 @@ public class ReviewService {
 
     @Transactional
     public Review deleteReview(Long reviewId, Long memberId) {
-        Long requestMemberId = memberLookupService.findByMemberId(memberId).getMemberId();
         Review review = reviewLookupService.findByReviewId(reviewId);
 
-        // TODO: 리뷰 작성 회원과 삭제 요청 회원이 같은지 확인, 아닌 경우 예외
+        // 리뷰 작성 회원과 삭제 요청 회원이 같은지 확인, 아닌 경우 예외 발생
+        if (!review.getMember().getMemberId().equals(memberId)) {
+            throw new NotReviewAuthorException(reviewId, memberId);
+        }
+
         repository.deleteByReviewId(reviewId);
         log.debug("review deleted: {}", review);
         return review;
